TileSet 0.1

+++ Base Game/Project_8_Solution/Source/ModulePlayer.h
@@ -5,6 +5,9 @@
 #include "Animation.h"
 #include "p2Point.h"
 
+//ESCENAS/NIVELES
+#include "SceneLevel1.h"
+
 struct SDL_Texture;
 struct Collider;
 
@@ -44,7 +47,7 @@ public:
 
 
 	// The speed in which we move the player (pixels per frame)
-	int speed = 1.0; //ERIC: Cambiada la velocidad a la que se mueve PacMam , preguntar profe
+	float speed = 1.0; //ERIC: Cambiada la velocidad a la que se mueve PacMam , preguntar profe
 
 	// The player spritesheet loaded into an SDL_Texture
 	SDL_Texture* texture = nullptr;
